Some restructuring for a general purpose mcmc
