+++ .gitignore
@@ -2,9 +2,9 @@
 /.settings
 *.aplx
 *.txt
-/build/
+**/build/
 /.pydevproject
-/application_generated_data_files/
-/reports/
+**/application_generated_data_files/
+**/reports/
 **/*.pyc
-results.*
+**/results.*
