+++ c_models/Makefile
@@ -1,0 +1,8 @@
+DIRS = mcmc_models mcmc_coordinator
+
+all: $(DIRS)
+	@for d in $(DIRS); do $(MAKE) -C $$d || exit $$?; done
+
+clean: $(DIRS)
+	@for d in $(DIRS); do $(MAKE) -C $$d GNU=$(GNU) clean || exit $$?; done
+
